+++ src/ddo/interfaces/EditableMetadata.ts
@@ -1,5 +1,4 @@
 import { EditableMetadataLinks } from './EditableMetadataLinks'
-import { ServicePrices } from './ServicePrices'
 
 export interface EditableMetadata {
   description?: string
